word embeddings with glove, lstm with pytorch.
Currently, one log line is viewed as a sentence.
Sentences are padded to have equal length.
Feed this to LSTM, use pad_packed_sequence() to mask 0-pad values.
Not clear if this is done correctly yet, because it seems, that pad_packed_sequence() is meant to be used on sequences,
but currently, the sentence is viewed as a 'whole', not a sequence of words. maybe different approach?