Use Pytorch's build-in Embedding layer as the first layer before the LSTM layers in the model.
The idea is to produce indices for every word in the alphabet. Pass these indicies to the embedding layer.
Given a sequence of length seq_length, the model is then supposed to predict the index of the next word, i.e. index.
Currently, this is not working properly yet, there's a problem with the dimension of the target and prediction of the model.